+++ packages/ui/CHANGELOG.md
@@ -11,6 +11,9 @@ The following is a curated list of changes in the Enact ui module, newest change
 ### Removed
 
 - `fit` prop from `ui/Transition` in favor of using `className`
+## [1.0.0-alpha.5] - 2016-12-16
+
+No changes.
 
 ## [1.0.0-alpha.4] - 2016-12-2